- adding a sort to get the months in order - moving `MemberCount` struct to models.reports
